refac init - something like this?
