+++ fb-assets.js
@@ -1,0 +1,12 @@
+module.exports = {
+  /**
+   * `packid`: {
+   *   `stickerid`: value
+   * }
+   */
+	facebookStickers: {
+		"227877430692340": {
+			"369239263222822": ":thumbsup:"
+		}
+	}
+}
